Add D2 auto-conversion for tags
Now when a tag is released, a new stage will also convert the tag to D2,
then it will commit the changes, create a new tag with a +d2 suffix,
push the changes and create a GitHub release for it.

This is done via the new beaver dlang d2-release command.

This is using a fork as a submodule because we'll make a test alpha
release just as part of testing beaver. The submodule should be updated
after the tests are finished.

* submodules/beaver v0.2.2(480e6f5)...v0.2.2+10(a8d6a35) (10 commits)
  > dlang: Add a command to release converted D2 tags
  > make-wrapper-dlang: Fix matrix row selection
  > Drop trusty support
  > Change default matrix row for make-wrapper-dlang
  > Add an make wrapper for dlang projects
  (...)